OneDrive/Desktop/Proctoring/backend/requirements.txt
opencv-python
mediapipe
mysql-connector-python
numpy
fastapi
uvicorn
ultralytics
torch
numpy
sounddevice

